+++ src/main/java/net/minecraft/world/storage/MapStorage.java
@@ -139,13 +139,13 @@ public class MapStorage {
         if (oshort == null) {
             oshort = (short) 0;
         } else {
-            oshort = (short) (oshort.shortValue() + 1);
+            oshort = (short) (oshort + 1);
         }
 
         this.idCounts.put(key, oshort);
 
         if (this.saveHandler == null) {
-            return oshort.shortValue();
+            return oshort;
         } else {
             try {
                 File file1 = this.saveHandler.getMapFileFromName("idcounts");
@@ -154,7 +154,7 @@ public class MapStorage {
                     NBTTagCompound nbttagcompound = new NBTTagCompound();
 
                     for (String s : this.idCounts.keySet()) {
-                        short short1 = this.idCounts.get(s).shortValue();
+                        short short1 = this.idCounts.get(s);
                         nbttagcompound.setShort(s, short1);
                     }
 
@@ -166,7 +166,7 @@ public class MapStorage {
                 exception.printStackTrace();
             }
 
-            return oshort.shortValue();
+            return oshort;
         }
     }
 }
