+++ src/main/java/net/optifine/DynamicLights.java
@@ -226,7 +226,7 @@ public class DynamicLights {
                         Integer integer = mapItemLightLevels.get(item);
 
                         if (integer != null) {
-                            return integer.intValue();
+                            return integer;
                         }
                     }
 
@@ -256,7 +256,7 @@ public class DynamicLights {
                     Integer integer = mapEntityLightLevels.get(entity.getClass());
 
                     if (integer != null) {
-                        return integer.intValue();
+                        return integer;
                     }
                 }
